\name{BackgroundOrganism}
\alias{BackgroundOrganism}
\docType{data}
\title{Probabilities of each nucleotide in the \emph{Homo sapiens} organism
according to Thakurta et al. }
\description{Probabilities of each nucleotide in the Homo sapiens organism
according to Thakurta et al.}
\usage{data(BackgroundOrganism)}
\format{
  \describe{
  \item{Prob}{Probability of each nucleotide}
  }
}
\source{
G. Guha Thakurta,Computational identification of transcriptional regulatory
elements in dna sequence, Nucleic Acids Res., vol. 34, pp. 3585-3598,
2006.}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{datasets}


