\name{CalculRedundancy}
\alias{CalculRedundancy}
\title{
CalculRedundancy: To calculate the redundancy 
}
\description{
This function calculates the redundancy (order) of a random variable X with N possible states.}
\usage{
CalculRedundancy(Factortrans, q,iicc)
}
\arguments{
  \item{Factortrans}{ A set of aligned nucleotide sequences}
  \item{q}{Renyi Order}
  \item{iicc}{A list options}
  
}

\author{
Joan Maynou <joan.maynouatupc.edu>}
\seealso{
redundancy, entropy.Shannon,entropy.Renyi, correction.redundancy, correction.entropy, correctionaprox
}

