\name{correction-class}
\alias{correction-class}
\docType{class}
\title{Correction for finite sample effect}
\description{Correction for the finite sample effect}
\format{
  \describe{
  \item{Herrorl}{Mean Entropy}
  \item{sderrorl}{Standard deviation Entropy}
  }
}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{class}