\name{DrosophilaDivergence}
\docType{data}
\alias{DrosophilaDivergence}
\title{A library of models for Divergence algorithm and \emph{Drosophila melanogaster} organism}
\description{
	A set of models for Divergence algorithm and \emph{Drosophila melanogaster} organism
}
\usage{DrosophilaDivergence}
\format{List}
\keyword{datasets}



