\name{DrosophilaEntropy}
\docType{data}
\alias{DrosophilaEntropy}
\title{A library of models for Entropy algorithm and \emph{Drosophila melanogaster} organism}
\description{
	A set of models for Entropy algorithm and \emph{Drosophila melanogaster} organism
}
\usage{DrosophilaEntropy}
\format{List}
\keyword{datasets}



