\name{DrosophilaQresiduals}
\docType{data}
\alias{DrosophilaQresiduals}
\title{A library of models for Qresiduals algorithm and \emph{Drosophila melanogaster} organism}
\description{
	A set of models for Qresiduals algorithm and \emph{Drosophila melanogaster} organism
}
\usage{DrosophilaQresiduals}
\format{List}
\keyword{datasets}



