\name{HomoDivergence}
\docType{data}
\alias{HomoDivergence}
\title{A library of models for Divergence algorithm and \emph{Homo sapiens} organism}
\description{
	A set of models for Divergence algorithm and \emph{Homo sapiens} organism
}
\usage{HomoDivergence}
\format{List}
\keyword{datasets}



