\name{HomoQresiduals}
\docType{data}
\alias{HomoQresiduals}
\title{A library of models for Qresiduals algorithm and \emph{Homo sapiens} organism}
\description{
	A set of models for Qresiduals algorithm and \emph{Homo sapiens} organism
}
\usage{HomoQresiduals}
\format{List}
\keyword{datasets}



