\name{iicc}
\alias{iicc}
\title{A set of initial conditions}
\description{A set of initial conditions for the MEET-package}
\docType{data}
\arguments{
  \item{TF}{ A set of nucleotide sequences}
  \item{nameTF}{Transcription Factor name}
  \item{seqin}{Candidate sequence}
  \item{alg}{A set of Multiple Sequence Alignment}
  \item{method}{A set of motif discovery algorithms}
  \item{mode}{Training or detection}
  \item{org}{Background organism}
  \item{vector}{Characteristic parameters methodfor leave-one-out cross-validation.}
  \item{num_motif}{Number of motif}
  \item{len_motif}{Length of motif}
  \item{direction}{Forward, reverse or both}
  \item{threshold}{P-value}
  \item{order}{Renyi Order}
  \item{model}{model}
  \item{position}{Binding site position for training}
  \item{mv}{missing values}
  \item{gapopen}{The gap open score}
  \item{maxiters}{Maximum number of iterations}
  \item{gapextend}{The gap extend score}
  \item{optionsFile}{Path for MEME/MAST, clustalw and muscle}
} 
\details{ Models (Entropy, Divergence and Qresiduals) for a set of TFBS. Homo sapiens:AP1,CREB1,E2F1,ELK1,ELK4,ESR1,ETS1,FEV,FOXD1,FOXO3,HLF,INSM1,Irf_2,MAX,MYC_MAX,MZF1_1_4,MZF1_5_13,NFATC2,NFE2L2,NFIL3,NFKB1,NF_kappaB,NKX3_1,NR3C1,NRF21,PPARG,Pax6,RORA_1,RORA_2,RREB1,REL,RELA,REST,RXRA_VDR,SP1,SPIB,SRF,SRY,STAT1,TLX1_NFIC,TAL1_TCF3,YY1,ZNF354C.
Rattus norvegicus:AP1,CREB1,Ddit3_Cebpa,FEV,Foxd3,Foxq1,Mafb,NFATC2,NF_kappaB,NR3C1,SP1.
Mus musculus:AP1,CREB1,Ddit3_Cebpa,FEV,Foxd3,Foxq1,Mafb,NFATC2,NF_kappaB,NR3C1,SP1.
Drosophila melanogaster:Abd_B,Antp,Awh,B_H1,B_H2,C15,CG11085,CG11294,CG11617,CG13424,CG15696,CG32105,CG32532,CG42234,CG4328,CG7056,CG9876,Deaf1,Dfd,Dll,Dr,E5,HGTX,Hmx,Kr,Lag1,Lim1,Lim3,NK7_1,Oct,PHDP,Pph13,Ptx1,Ptx1,Rx,Six4,Vsx1,Vsx2,a1,abd_A,achi,ap,ara,bcd,brk,bsh,btd,btn,cad,caup,ct,dl_1,ems,en,eve,exd,exex,fkh,ftz,gt,h,hbn,hkb,hth,ind,inv,kni,lab,lbe,mirr,nub,oc,odd,onecut,opa,otp,pan,pb,ro,sd,slbo,slou,slp1,sna,so,tin,tll,tup,twi,unc_4,unpg,usp,vis,vnd,vvl,zen,Gsc,dl_2,ttk,repo,Scr}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}


