\name{Model-class}
\alias{Model-class}
\docType{class}
\title{A set of Models for the detection}
\description{A set of models for the detection}
\format{
  \describe{
  \item{Model}{A List of the models}
  }
}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{class}

