\name{motif.mast}
\alias{motif.mast}
\title{MEME format to training matrix}
\description{Given a set of DNA sequences in fasta format, this function converts this set of sequences in a training matrix}
\usage{
motif.mast(input,Factortrans,k,m)
}
\arguments{
\item{input}{a set of DNA seqeunces in fasta format}
\item{Factortrans}{Transcription Factor Binding Sites}
\item{k}{Length of the candidate sequence}
\item{m}{Binnding Sites sequences that is extracted from training matrix}
}
\author{
Joan Maynou <joan.maynouatupc.edu> and Erola Pairo <epeiroatibec.pcb.ub.es>
}


