\name{MusDivergence}
\docType{data}
\alias{MusDivergence}
\title{A library of models for Divergence algorithm and \emph{Mus musculus} organism}
\description{
	A set of models for Divergence algorithm and \emph{Mus musculus} organism}
\usage{MusDivergence}
\format{List}
\keyword{datasets}



