\name{MusQresiduals}
\docType{data}
\alias{MusQresiduals}
\title{A library of models for Qresiduals algorithm and \emph{Mus musculus} organism}
\description{
	A set of models for Qresiduals algorithm and \emph{Mus musculus} organism}
\usage{MusQresiduals}
\format{List}
\keyword{datasets}



