\name{numericalTF}
\docType{data}
\alias{NumericalTF}
\title{Matrix of a numerical motif}
\description{
Matrix with a numerical DNA motif.  
}
\usage{numericalTF}
\format{Matrix}
\keyword{datasets}
\references{
Silverman, B.D., and Linske, R.: A measure of DNA periodicity, Journal of Theoretical Biology 118, volume 118,p. 295, 1986
}


