\name{organism}
\alias{organism}
\docType{data}
\title{Probability for each nucleotide according to different
organism}
\description{Probability for each nucleotide according to different
organism: \emph{Saccharomyces cerevisiae}, \emph{Drosophila
melanogaster}, \emph{Homo sapiens}, \emph{Escherichia coli},
\emph{Caenorhabditis elegans}, \emph{Plasmodium falciparum}. \emph{Streptomyces coelicolor}, \emph{Gallus gallys}, \emph{Mus musculus} and \emph{Rattus norvegicus}}

\usage{data(organism)}

\format{
  \describe{
  \item{organism}{Table of probability for each nucleotide according to different organism}
  }
}

\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{datasets}