\name{probability.couple}
\alias{probability.couple}
\title{
Background joint probability}
\description{
Background joint probability
}
\usage{
probability.couple(Prob)
}
\arguments{
	\item{Prob}{Probability of each base within genome}}
\author{Joan Maynou <joan.maynou@upc.edu>}


\seealso{
probability, joint.probability, probIdem, probtransicio
}
\examples{
data(BackgroundOrganism)
probability.couple(Prob)
}

