\name{RattusDivergence}
\docType{data}
\alias{RattusDivergence}
\title{A library of models for Divergence algorithm and \emph{Rattus norvegicus} organism}
\description{
	A set of models for Divergence algorithm and \emph{Rattus norvegicus} organism}
\usage{RattusDivergence}
\format{List}
\keyword{datasets}



