\name{RattusEntropy}
\docType{data}
\alias{RattusEntropy}
\title{A library of models for Entropy algorithm and \emph{Rattus norvegicus} organism}
\description{
	A set of models for Entropy algorithm and \emph{Rattus norvegicus} organism}
\usage{RattusEntropy}
\format{List}
\keyword{datasets}



