\name{RattusQresiduals}
\docType{data}
\alias{RattusQresiduals}
\title{A library of models for Qresiduals algorithm and \emph{Rattus norvegicus} organism}
\description{
	A set of models for Qresiduals algorithm and \emph{Rattus norvegicus} organism}
\usage{RattusQresiduals}
\format{List}
\keyword{datasets}



