\name{read.mast}
\alias{read.mast}
\title{Read output mast}
\description{This function reads the output of MAST}
\usage{
read.mast(input, Factortrans, k, m)
}

\arguments{
  \item{input}{ Output MAST}
  \item{Factortrans}{A set of aligned nucleotide sequences}
  \item{k}{Length DNA candidate sequence}
  \item{m}{Number of TFBS sequence used on leave-one-out cross-validation}
}

\author{
Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>}


