\name{RenyiOrder}
\docType{data}
\alias{Order}
\title{Renyi entropy and Divergence order}
\description{
parameter q, needed to calculate the Renyi Entropy and the Divergence
}
\usage{RenyiOrder}
\format{number}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, Applied bioinformatics for the identification of regulatory elements, Nat . Rev. Genet. vol. 5, no. 4, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


