\name{scoreMDscan}
\alias{scoreMDscan}
\title{
Output MDscan method
}
\description{
This function writes the output MDscan method. The output contains the next fields:sequence, direction, score MDscan.}
\usage{
scoreMDscan(input, k, matriu, direction)
}
\arguments{
  \item{input}{Score MDscan}
  \item{k}{Length DNA sequence}
  \item{matriu}{A set of aligned nucleotide sequence}
  \item{direction}{Direction of DNA sequence}
}
\details{
Output's run.read.MDscan is input's scoreMDscan}

\author{
Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>
}


\seealso{
run.read.MDscan
}

