\name{Sequence}
\alias{Sequence}
\docType{data}
\title{A sequence with binding evidence.}

\description{A sequence with binding evidence}

\usage{data(Sequence)}

\format{
  \describe{
  \item{Sequence}{Binding sites sequence}
  }
}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{datasets}

