\name{TFlogodds}
\alias{TFlogodds}
\docType{data}
\title{Logodds matrix}
\description{Logodds matrix}

\usage{data(TFlogodds)}

\format{
  \describe{
  \item{TFlogodds}{Log odds for a transcription factor binding sites}
  } 
}

\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{datasets}