\name{TranscriptionFactor}
\alias{TranscriptionFactor}
\docType{data}
\title{A set of aligned binding sites sequences}
\description{Binding sites sequences}

\usage{data(TranscriptionFactor)}

\format{
  \describe{
  \item{TranscriptionFactor}{A set of aligned binding sites sequences}
  }
}
\author{Erola Pairo <epeiroatibec.pcb.ub.es> and Joan Maynou <joan.maynouatupc.edu>.}
\keyword{datasets}
