\name{BackgroundOrganism}
\docType{data}
\alias{BackgroundOrganism}
\title{Frequency vector}
\description{
	Background frequency of the each letter at the same position
	from \emph{Saccharomyces cerevisiae}
}
\usage{BackgroundOrganism}
\format{Vector}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, “Applied bioinformatics for the identification of regulatory elements,” Nat . Rev. Genet. vol. 5, no. 4, pp. 276–287, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


