\name{iicc}
\docType{data}
\alias{initial conditions}
\title{Initial conditions for the MEET package}
\description{
list with the basic initial conditions needed to run the MEET package. e.g.: the alignment method, the algorithm, the training mode, etc  
}
\usage{iicc}
\format{List}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin,Applied bioinformatics for the identification of regulatory elements, Nat . Rev. Genet. vol. 5, no. 4, pp. 276, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


