\name{MImemory}
\alias{MImemory}
\title{Library of PredictDivergence values
}
\description{ This function calculates divergence values from all combinations possibles of a couple of nucleotides given a matrix of TFBS sequences. The divergence values is calculated only by TFBS positions correlated. }
\usage{
MImemory(iicc, training.set)
}

\arguments{
\item{iicc}{A set of inicial conditions for the MEET-package}
  \item{training.set}{ A set of nucleotide sequences}
}

\author{
Joan Maynou <joan.maynou@upc.edu>
}

\seealso{
MIread}


