\name{Model}
\docType{class}
\alias{Model}
\title{Creates a model class}
\description{ This function creates a Model class and it should not be
used by the user}
\section{Slots}{
\describe{
	\item{model}{model}
	}
}
\author{
Joan Maynou <joan.maynou@upc.edu> and Erola Pairo  <epairo@ibecbarcelona.eu>
}


