\name{Model}
\alias{Model}
\title{Creates a model class}
\description{ This function creates a Model class and it should not be used by the user}
\usage{
Models(iicc)
}

\arguments{
\item{ } defines the class
}

\author{
Joan Maynou <joan.maynou@upc.edu> and Erola Pairó  <epairo@ibecbarcelona.eu>
}
\examples{
Models()
}

