\name{numericalTF}
\docType{data}
\alias{NumericalTF}
\title{Matrix of a numerical motif}
\description{
Matrix with a numerical DNA motif.  
}
\usage{numericalTF}
\format{Matrix}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, “Applied bioinformatics for the identification of regulatory elements,” Nat . Rev. Genet. vol. 5, no. 4, pp. 276–287, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


