\name{organism}
\docType{data}
\alias{organism}
\title{Matrix of background frequencies}
\description{
	Background frequencies of each of the nucleotides for all the available organisms
}
\usage{organism}
\format{Matrix}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, Applied bioinformatics for the identification of regulatory elements,Nat . Rev. Genet. vol. 5, no. 4, pp. 276, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


