\name{TFlogodds}
\docType{data}
\alias{TFlogodds}
\title{Logodds for a motif PSSM}
\description{
Logodds of the PSSM of the ROX1 motif, calculated using information theory
}
\usage{TFlogodds}
\format{Matrix}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, Applied bioinformatics for the identification of regulatory elements, Nat . Rev. Genet. vol. 5, no. 4, pp. 276, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


