\name{TranscriptionFactor}
\docType{data}
\alias{TranscriptionFactor}
\title{Transcription Factor Sequences}
\description{
Aligned binding sites for a transciption Factor
}
\usage{TranscriptionFactor}
\format{data frame}
\keyword{datasets}
\references{
W. W. Wasserman and A. Sandelin, Applied bioinformatics for the identification of regulatory elements, Nat . Rev. Genet. vol. 5, no. 4, Apr. 2004 [Online].Available: http://dx.doi.org/10.1038/ nrg1315
}


